+++ Sorter.BigFiles.Generator/Sorter.BigFiles.Generator/ConfigOptions.cs
@@ -6,7 +6,7 @@
 
         public string OutputFilesDirectory { get; set; } = string.Empty;
         public string OutputFileName { get; set; } = string.Empty;        
-        public int OutputFileSizeInMBs { get; set; } = 100;
+        public long OutputFileSizeInMBs { get; set; } = 100;
         public string OutputStringFormat { get; set; } = string.Empty;
 
         public string InputStingsFileName { get; set; } = string.Empty;
